refactor: extract lib tests to sibling file
